create subclass for db and csv
