+++ output/CSVOutputter.py
@@ -1,27 +1,21 @@
-"""A file containing the Outputter class and its methods.
+"""A specific class to output information to a csv file
 Authors: Bruno Lerner, Doria Philo, Yuri Kaz"""
 
 import csv
 import os
-from config_file import HEADER
+from config_file import HEADER, OUTPUT_FILENAME
+from output import outputter
 
 
-class Outputter:
+class CSVOutputter(outputter.Outputter):
     """Used to output the content to a specified file."""
 
-    def __init__(self, output_filename, file_type, quotes_objects):
+    def __init__(self, file_type, quotes_objects):
         self.file_type = file_type
-        self.filename = output_filename + '.' + self.file_type
+        self.filename = OUTPUT_FILENAME + '.' + self.file_type
         self.quotes_objects = quotes_objects
 
-    def write_file(self):
-        """Writes the info of each entry to a file according to the
-        specified file type."""
-        if self.file_type == 'csv':
-            self._write_to_csv()
-        # Other relevant file types can be added here in the future
-
-    def _write_to_csv(self):
+    def create_output(self):
         """Writes the info of each entry to a csv file."""
         with open(self.filename, 'a', newline='', encoding='utf-8') as \
                 csv_file:
